refactor: address review on set-cookie handling
